fix(cursor): preserve crosshair pixels on aligned canvas
