+++ src/shot_window_internal.cpp
@@ -26,43 +26,6 @@ QString magnifierScaleText(qreal scale)
     return QStringLiteral("%1x").arg(QString::number(clampedMagnifierScale(scale), 'f', 2));
 }
 
-/// @brief Creates the high-contrast cross cursor used during capture.
-/// @return Cursor used for selection and annotation drawing modes.
-QCursor captureCrossCursor()
-{
-    // Use 64x64 to guarantee 256-byte row pitch. 
-    // This fixes hardware cursor tearing/stride bugs on Hyprland/wlroots DRM backends
-    constexpr int canvasSize = 64;
-    constexpr int crossSize = 32;
-
-    QPixmap pixmap(canvasSize, canvasSize);
-    pixmap.fill(Qt::transparent);
-
-    QPainter painter(&pixmap);
-    painter.setRenderHint(QPainter::Antialiasing, false);
-    
-    // Calculate drawing boundaries dynamically
-    constexpr int startPos = (canvasSize - crossSize) / 2;
-    constexpr int endPos = startPos + crossSize - 1;
-    constexpr int centerPos = (canvasSize - 1) / 2;
-
-    painter.setPen(QPen(QColor(15, 23, 42, 235), 5, Qt::SolidLine, Qt::SquareCap));
-    painter.drawLine(centerPos, startPos, centerPos, endPos);
-    painter.drawLine(startPos, centerPos, endPos, centerPos);
-    
-    painter.setPen(QPen(QColor(255, 255, 255, 245), 3, Qt::SolidLine, Qt::SquareCap));
-    painter.drawLine(centerPos, startPos, centerPos, endPos);
-    painter.drawLine(startPos, centerPos, endPos, centerPos);
-    
-    painter.setPen(QPen(QColor(45, 212, 191, 255), 1, Qt::SolidLine, Qt::SquareCap));
-    painter.drawLine(centerPos, startPos, centerPos, endPos);
-    painter.drawLine(startPos, centerPos, endPos, centerPos);
-    
-    painter.end();
-
-    return QCursor(pixmap, centerPos, centerPos);
-}
-
 qreal normalizedRotationDegrees(qreal degrees)
 {
     degrees = std::fmod(degrees, 360.0);
@@ -270,238 +233,6 @@ qreal imageNavigationWheelFactor(const QWheelEvent *event)
     }
 
     return 1.0;
-}
-
-/// @brief A single resampling sample representing an input pixel weight.
-struct AxisSample {
-    /// @brief The input index of the sample.
-    int index = 0;
-    /// @brief The weight contribution of the sample.
-    double weight = 0.0;
-};
-
-/// @brief Resampling lookup table for a single coordinate axis.
-struct AxisTable {
-    /// @brief Precalculated kernel samples for each output coordinate.
-    std::vector<std::vector<AxisSample>> samples;
-    /// @brief The lowest input coordinate index referenced in the table.
-    int first = 0;
-    /// @brief The highest input coordinate index referenced in the table.
-    int last = -1;
-};
-
-/// @brief Evaluates a sharpening filter kernel.
-/// @param distance The distance from the center.
-/// @return The filter weight at the given distance.
-double sharpKernel(double distance)
-{
-    const double x = std::abs(distance);
-    if (x > kSharpKernelRadius) {
-        return 0.0;
-    }
-    if (x <= 0.5) {
-        return 17.0 / 16.0 - 7.0 / 4.0 * x * x;
-    }
-    if (x <= 1.5) {
-        return x * x - 11.0 / 4.0 * x + 7.0 / 4.0;
-    }
-    return -1.0 / 8.0 * x * x + 5.0 / 8.0 * x - 25.0 / 32.0;
-}
-
-/// @brief Builds an AxisTable for 1D image scaling.
-/// @param inputSize The size of the input axis.
-/// @param outputSize The size of the output axis.
-/// @param sourceStart The starting coordinate in the source coordinate system.
-/// @param scale The scaling factor.
-/// @return The populated AxisTable.
-AxisTable buildAxisTable(int inputSize, int outputSize, double sourceStart, double scale)
-{
-    AxisTable table;
-    table.samples.resize(outputSize);
-    table.first = inputSize;
-
-    const double filterScale = std::min(scale, 1.0);
-    const double radius = kSharpKernelRadius / filterScale;
-    for (int output = 0; output < outputSize; ++output) {
-        const double center = sourceStart + (static_cast<double>(output) + 0.5) / scale - 0.5;
-        const int first = std::max(0, static_cast<int>(std::floor(center - radius)));
-        const int last = std::min(inputSize - 1, static_cast<int>(std::ceil(center + radius)));
-
-        double sum = 0.0;
-        std::vector<AxisSample> samples;
-        samples.reserve(std::max(0, last - first + 1));
-        for (int input = first; input <= last; ++input) {
-            const double weight = sharpKernel((static_cast<double>(input) - center) * filterScale);
-            if (weight == 0.0) {
-                continue;
-            }
-            samples.push_back({input, weight});
-            sum += weight;
-        }
-
-        if (samples.empty() || qFuzzyIsNull(sum)) {
-            const int nearest = std::clamp(static_cast<int>(std::round(center)), 0, inputSize - 1);
-            samples.push_back({nearest, 1.0});
-            table.first = std::min(table.first, nearest);
-            table.last = std::max(table.last, nearest);
-        } else {
-            for (AxisSample &sample : samples) {
-                sample.weight /= sum;
-                table.first = std::min(table.first, sample.index);
-                table.last = std::max(table.last, sample.index);
-            }
-        }
-
-        table.samples[output] = std::move(samples);
-    }
-
-    if (table.first > table.last) {
-        table.first = 0;
-        table.last = 0;
-    }
-    return table;
-}
-
-/// @brief Converts a weighted double value into a byte.
-/// @param value The double value to convert.
-/// @return The converted integer value clamped between 0 and 255.
-int byteFromWeightedSum(double value)
-{
-    return std::clamp(static_cast<int>(std::lround(value)), 0, 255);
-}
-
-/// @brief Blends and packs red, green, blue, and alpha components into a premultiplied QRgb value.
-/// @param red The red color component.
-/// @param green The green color component.
-/// @param blue The blue color component.
-/// @param alpha The alpha opacity component.
-/// @return The premultiplied QRgb value.
-QRgb premultipliedPixel(double red, double green, double blue, double alpha)
-{
-    const int a = byteFromWeightedSum(alpha);
-    const int r = std::min(byteFromWeightedSum(red), a);
-    const int g = std::min(byteFromWeightedSum(green), a);
-    const int b = std::min(byteFromWeightedSum(blue), a);
-    return qRgba(r, g, b, a);
-}
-
-template <typename Function>
-void forRowRanges(int rowCount, Function function)
-{
-    if (rowCount <= 0) {
-        return;
-    }
-
-    /// @brief The ideal number of concurrent threads to use.
-    const int idealThreads = std::max(1, QThread::idealThreadCount());
-    /// @brief The number of threads selected for processing.
-    const int threadCount = std::clamp(rowCount / kMinSharpRowsPerThread, 1, idealThreads);
-    if (threadCount == 1) {
-        function(0, rowCount);
-        return;
-    }
-
-    /// @brief The ranges of rows to be processed by each thread.
-    std::vector<std::pair<int, int>> ranges;
-    ranges.reserve(threadCount);
-    /// @brief The number of rows allocated to each thread.
-    const int rowsPerThread = rowCount / threadCount;
-    /// @brief The starting row index of the current range.
-    int begin = 0;
-    /// @brief The loop counter for iterating thread ranges.
-    for (int i = 0; i < threadCount; ++i) {
-        /// @brief The ending row index of the current range.
-        const int end = i == threadCount - 1 ? rowCount : begin + rowsPerThread;
-        ranges.push_back({begin, end});
-        begin = end;
-    }
-
-    QtConcurrent::blockingMap(ranges, [&function](const std::pair<int, int> &range) {
-        function(range.first, range.second);
-    });
-}
-
-QImage renderSharpViewport(const QImage &source, const QRectF &sourceRect, const QSize &targetSize)
-{
-    if (source.isNull() || sourceRect.isEmpty() || targetSize.isEmpty()) {
-        return {};
-    }
-
-    const QImage src = source.format() == QImage::Format_ARGB32_Premultiplied
-        ? source
-        : source.convertToFormat(QImage::Format_ARGB32_Premultiplied);
-    const QRectF sourceBounds(QPointF(0.0, 0.0), QSizeF(src.size()));
-    const QRectF clampedSourceRect = sourceRect.intersected(sourceBounds);
-    if (clampedSourceRect.isEmpty()) {
-        return {};
-    }
-
-    const double scaleX = static_cast<double>(targetSize.width()) / clampedSourceRect.width();
-    const double scaleY = static_cast<double>(targetSize.height()) / clampedSourceRect.height();
-    const AxisTable xTable =
-        buildAxisTable(src.width(), targetSize.width(), clampedSourceRect.left(), scaleX);
-    const AxisTable yTable =
-        buildAxisTable(src.height(), targetSize.height(), clampedSourceRect.top(), scaleY);
-
-    const int intermediateHeight = yTable.last - yTable.first + 1;
-    QImage horizontal(targetSize.width(), intermediateHeight, QImage::Format_ARGB32_Premultiplied);
-    const uchar *sourceBits = src.constBits();
-    const int sourceStride = src.bytesPerLine();
-    uchar *horizontalBits = horizontal.bits();
-    const int horizontalStride = horizontal.bytesPerLine();
-
-    forRowRanges(intermediateHeight, [&](int begin, int end) {
-        for (int row = begin; row < end; ++row) {
-            const int y = yTable.first + row;
-            const auto *sourceLine =
-                reinterpret_cast<const QRgb *>(sourceBits + y * sourceStride);
-            auto *targetLine =
-                reinterpret_cast<QRgb *>(horizontalBits + row * horizontalStride);
-            for (int x = 0; x < targetSize.width(); ++x) {
-                double a = 0.0;
-                double r = 0.0;
-                double g = 0.0;
-                double b = 0.0;
-                for (const AxisSample &sample : xTable.samples[x]) {
-                    const QRgb pixel = sourceLine[sample.index];
-                    a += sample.weight * qAlpha(pixel);
-                    r += sample.weight * qRed(pixel);
-                    g += sample.weight * qGreen(pixel);
-                    b += sample.weight * qBlue(pixel);
-                }
-                targetLine[x] = premultipliedPixel(r, g, b, a);
-            }
-        }
-    });
-
-    QImage target(targetSize, QImage::Format_ARGB32_Premultiplied);
-    const uchar *horizontalReadBits = horizontal.constBits();
-    const int horizontalReadStride = horizontal.bytesPerLine();
-    uchar *targetBits = target.bits();
-    const int targetStride = target.bytesPerLine();
-
-    forRowRanges(targetSize.height(), [&](int begin, int end) {
-        for (int y = begin; y < end; ++y) {
-            auto *targetLine = reinterpret_cast<QRgb *>(targetBits + y * targetStride);
-            for (int x = 0; x < targetSize.width(); ++x) {
-                double a = 0.0;
-                double r = 0.0;
-                double g = 0.0;
-                double b = 0.0;
-                for (const AxisSample &sample : yTable.samples[y]) {
-                    const auto *sourceLine = reinterpret_cast<const QRgb *>(
-                        horizontalReadBits + (sample.index - yTable.first) * horizontalReadStride);
-                    const QRgb pixel = sourceLine[x];
-                    a += sample.weight * qAlpha(pixel);
-                    r += sample.weight * qRed(pixel);
-                    g += sample.weight * qGreen(pixel);
-                    b += sample.weight * qBlue(pixel);
-                }
-                targetLine[x] = premultipliedPixel(r, g, b, a);
-            }
-        }
-    });
-    return target;
 }
 
 // Stylesheets, palette presets, action names, and toolbar icons now live in
